Added exception
I added an exception to each of the codes for what happens when the
supplied argument refers to a file that doesn't exist, or in the case of
python_test_3, isn't a number.
This probably wasn't necessary for the purposes of the codes, but I like
to know that codes I write cover their bases.